requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
jinja2>=3.1.0
assemblyai
murf
websockets>=11.0
requests>=2.31.0
pytz>=2023.3

google-generativeai>=0.5.0
